archive_env fixture added to integration conftest
