+++ tests/integration/test_archives.py
@@ -18,11 +18,9 @@ def test_repo_list(qapp, qtbot):
     """Test that the archives are created and repo list is populated correctly"""
     main = qapp.main_window
     tab = main.archiveTab
-
     main.tabWidget.setCurrentIndex(3)
     tab.refresh_archive_list()
     qtbot.waitUntil(lambda: not tab.bCheck.isEnabled(), **pytest._wait_defaults)
-
     assert not tab.bCheck.isEnabled()
 
     qtbot.waitUntil(lambda: 'Refreshing archives done.' in main.progressText.text(), **pytest._wait_defaults)
@@ -31,30 +29,18 @@ def test_repo_list(qapp, qtbot):
     assert tab.bCheck.isEnabled()
 
 
-def test_repo_prune(qapp, qtbot):
+def test_repo_prune(qapp, qtbot, archive_env):
     """Test for archive pruning"""
-    main = qapp.main_window
-    tab = main.archiveTab
-
-    main.tabWidget.setCurrentIndex(3)
-    tab.refresh_archive_list()
-    qtbot.waitUntil(lambda: tab.archiveTable.rowCount() > 0, **pytest._wait_defaults)
-
+    main, tab = archive_env
     qtbot.mouseClick(tab.bPrune, QtCore.Qt.MouseButton.LeftButton)
     qtbot.waitUntil(lambda: 'Pruning old archives' in main.progressText.text(), **pytest._wait_defaults)
     qtbot.waitUntil(lambda: 'Refreshing archives done.' in main.progressText.text(), **pytest._wait_defaults)
 
 
 @pytest.mark.min_borg_version('1.2.0a1')
-def test_repo_compact(qapp, qtbot):
+def test_repo_compact(qapp, qtbot, archive_env):
     """Test for archive compaction"""
-    main = qapp.main_window
-    tab = main.archiveTab
-
-    main.tabWidget.setCurrentIndex(3)
-    tab.refresh_archive_list()
-    qtbot.waitUntil(lambda: tab.archiveTable.rowCount() > 0, **pytest._wait_defaults)
-
+    main, tab = archive_env
     qtbot.waitUntil(lambda: tab.compactButton.isEnabled(), **pytest._wait_defaults)
     assert tab.compactButton.isEnabled()
 
@@ -62,14 +48,9 @@ def test_repo_compact(qapp, qtbot):
     qtbot.waitUntil(lambda: 'compaction freed about' in main.logText.text().lower(), **pytest._wait_defaults)
 
 
-def test_check(qapp, qtbot):
+def test_check(qapp, qtbot, archive_env):
     """Test for archive consistency check"""
-    main = qapp.main_window
-    tab = main.archiveTab
-
-    main.tabWidget.setCurrentIndex(3)
-    tab.refresh_archive_list()
-    qtbot.waitUntil(lambda: tab.archiveTable.rowCount() > 0, **pytest._wait_defaults)
+    main, tab = archive_env
 
     qapp.check_failed_event.disconnect()
 
@@ -81,7 +62,7 @@ def test_check(qapp, qtbot):
 
 
 @pytest.mark.skipif(sys.platform == 'darwin', reason="Macos fuse support is uncertain")
-def test_mount(qapp, qtbot, monkeypatch, choose_file_dialog, tmpdir):
+def test_mount(qapp, qtbot, monkeypatch, choose_file_dialog, tmpdir, archive_env):
     """Test for archive mounting and unmounting"""
 
     def psutil_disk_partitions(**kwargs):
@@ -91,12 +72,7 @@ def test_mount(qapp, qtbot, monkeypatch, choose_file_dialog, tmpdir):
     monkeypatch.setattr(psutil, "disk_partitions", psutil_disk_partitions)
     monkeypatch.setattr(vorta.views.archive_tab, "choose_file_dialog", choose_file_dialog)
 
-    main = qapp.main_window
-    tab = main.archiveTab
-
-    main.tabWidget.setCurrentIndex(3)
-    tab.refresh_archive_list()
-    qtbot.waitUntil(lambda: tab.archiveTable.rowCount() > 0, **pytest._wait_defaults)
+    main, tab = archive_env
     tab.archiveTable.selectRow(0)
 
     qtbot.waitUntil(lambda: tab.bMountRepo.isEnabled(), **pytest._wait_defaults)
@@ -114,14 +90,9 @@ def test_mount(qapp, qtbot, monkeypatch, choose_file_dialog, tmpdir):
     qtbot.waitUntil(lambda: tab.mountErrors.text().startswith('Un-mounted successfully.'), **pytest._wait_defaults)
 
 
-def test_archive_extract(qapp, qtbot, monkeypatch, choose_file_dialog, tmpdir):
+def test_archive_extract(qapp, qtbot, monkeypatch, choose_file_dialog, tmpdir, archive_env):
     """Test for archive extraction"""
-    main = qapp.main_window
-    tab = main.archiveTab
-
-    main.tabWidget.setCurrentIndex(3)
-    tab.refresh_archive_list()
-    qtbot.waitUntil(lambda: tab.archiveTable.rowCount() > 0, **pytest._wait_defaults)
+    main, tab = archive_env
 
     tab.archiveTable.selectRow(2)
     tab.extract_action()
@@ -139,14 +110,9 @@ def test_archive_extract(qapp, qtbot, monkeypatch, choose_file_dialog, tmpdir):
     assert [item.basename for item in tmpdir.listdir()] == ['private' if sys.platform == 'darwin' else 'tmp']
 
 
-def test_archive_delete(qapp, qtbot, mocker):
+def test_archive_delete(qapp, qtbot, mocker, archive_env):
     """Test for archive deletion"""
-    main = qapp.main_window
-    tab = main.archiveTab
-
-    main.tabWidget.setCurrentIndex(3)
-    tab.refresh_archive_list()
-    qtbot.waitUntil(lambda: tab.archiveTable.rowCount() > 0, **pytest._wait_defaults)
+    main, tab = archive_env
 
     archivesCount = tab.archiveTable.rowCount()
 
@@ -160,14 +126,9 @@ def test_archive_delete(qapp, qtbot, mocker):
     assert tab.archiveTable.rowCount() == archivesCount - 1
 
 
-def test_archive_rename(qapp, qtbot, mocker):
+def test_archive_rename(qapp, qtbot, mocker, archive_env):
     """Test for archive renaming"""
-    main = qapp.main_window
-    tab = main.archiveTab
-
-    main.tabWidget.setCurrentIndex(3)
-    tab.refresh_archive_list()
-    qtbot.waitUntil(lambda: tab.archiveTable.rowCount() > 0, **pytest._wait_defaults)
+    main, tab = archive_env
 
     tab.archiveTable.selectRow(0)
     new_archive_name = 'idf89d8f9d8fd98'
